changes as per recent feedback
