--- photo_fetcher/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Photo Fetcher</title>
-    <link rel="stylesheet" href="index.css" />
-</head>
-
-<body>
-    <header>
-        <h1>Photo Fetcher</h1>
-    </header>
-    <nav>
-        <label class="switch">
-            <input type="checkbox" id="toggle-switch"/>
-            <span class="slider"></span>
-            Make photos grayscale
-          </label>
-          <button class="fetch-button" id="fetch-photos">Fetch New Photos</button>
-    </nav>
-    <section aria-label="Photo gallery" class="photo-gallery" id="photo-gallery">
-    </section>
-    <footer>
-        <button class="fetch-button stretch" id="fetch-more-photos">More Photos</button>
-    </footer>
-</body>
-<script src="./index.js"></script>
-</html>
