+++ src/components/PhotoCard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import '../../css/app.css';
+import '../css/app.css';
 
 function PhotoCard({ photo, grayscale }) {
     return (
